+++ crates/milli/src/fields_ids_map/global.rs
@@ -123,7 +123,7 @@ impl MutFieldIdMapper for GlobalFieldsIdsMap<'_> {
     fn insert(&mut self, name: &str) -> Option<FieldId> {
         self.id_or_insert(name)
     }
-    fn len(&mut self) -> i32 {
-        GlobalFieldsIdsMap::len(self) as i32
+    fn len(&mut self) -> usize {
+        GlobalFieldsIdsMap::len(self) as usize
     }
 }
